separe los componentes para top bar y bottom nav
