+++ app/(tabs)/home.tsx
@@ -1,7 +1,9 @@
 // HomeScreen.tsx
 import React from 'react';
 import { SafeAreaView, ScrollView, View, StyleSheet } from 'react-native';
-import { TopBar, WeatherSection, Categories, RecentSection, Welcome, BottomNav } from '../components/Dashboard';
+import { WeatherSection, Categories, RecentSection, Welcome } from '../components/Dashboard';
+import { TopBar } from '../components/Dashboard_header';
+import { BottomNav } from '../components/Dashboard_Footer';
 
 const HomeScreen: React.FC = () => {
     return (
